# vps/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Retiradas API
  version: 1.0.0
  description: |
    Documentacao REST da API do sistema Retiradas.

    A API roda em `https://retiradas.tech/api` e usa JWT para rotas autenticadas.
    Rotas de escrita autenticadas tambem exigem CSRF via cookie/header, conforme o cliente web do sistema.

    Esta documentacao foi criada para orientar integracoes internas e parceiros tecnicos.
  contact:
    name: Retiradas.tech
servers:
  - url: https://retiradas.tech/api
    description: Producao
...
security:
  - bearerAuth: []
paths:
...
  /documents:
    get:
      tags: [Documents]
      summary: Lista documentos por collection
      parameters:
        - in: query
          name: collection
          required: true
          schema:
            type: string
          example: usuarios
        - in: query
          name: limit
          schema:
            type: integer
            default: 100
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Lista de documentos
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/AppDocument"
    post:
      tags: [Documents]
      summary: Cria documento em uma collection
      security:
        - bearerAuth: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                collection:
                  type: string
                  example: integracoes
                data:
                  type: object
                  additionalProperties: true
      responses:
        "200":
          description: Documento criado
  /documents/{path}:
    get:
      tags: [Documents]
      summary: Busca documento por caminho
      parameters:
        - $ref: "#/components/parameters/DocumentPath"
      responses:
        "200":
          description: Documento encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppDocument"
    put:
      tags: [Documents]
      summary: Atualiza documento por caminho
      security:
        - bearerAuth: []
          csrfToken: []
      parameters:
        - $ref: "#/components/parameters/DocumentPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    delete:
      tags: [Documents]
      summary: Remove documento por caminho
      security:
        - bearerAuth: []
          csrfToken: []
      parameters:
        - $ref: "#/components/parameters/DocumentPath"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
...
  /documentos/submissions:
    get:
      tags: [Documentos]
      summary: Lista envios de documentos
      responses:
        "200":
          description: Lista de envios
    post:
      tags: [Documentos]
      summary: Cria envio de documentos PDF para aprovacao
      security:
        - bearerAuth: []
          csrfToken: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                empresaId:
                  type: string
                mesReferencia:
                  type: string
                  example: "2026-08"
                files:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        "200":
          description: Envio criado
  /documentos/submissions/{id}/review:
    patch:
      tags: [Documentos]
      summary: Aprova ou reprova documentos
      security:
        - bearerAuth: []
          csrfToken: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/Ok"
...
  /documentos/folders/{empresaId}:
    post:
      tags: [Documentos]
      summary: Garante pasta da empresa no Google Drive
      security:
        - bearerAuth: []
          csrfToken: []
      parameters:
        - in: path
          name: empresaId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Pasta criada ou encontrada
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    csrfToken:
      type: apiKey
      in: header
      name: x-csrf-token
  parameters:
    DocumentPath:
      in: path
      name: path
      required: true
      schema:
        type: string
      description: Caminho completo do documento. Exemplo `usuarios/uid`.
...
  responses:
    Ok:
      description: Operacao realizada
      content:
        application/json:
          schema:
            type: object
            properties:
              ok:
                type: boolean
                example: true
...
  schemas:
...
    AppDocument:
      type: object
      properties:
        path:
          type: string
        collectionPath:
          type: string
        documentId:
          type: string
        parentPath:
          type: string
          nullable: true
        data:
          type: object
          additionalProperties: true
        updatedAt:
          type: string
          format: date-time
...